agent/pyproject.toml
[project]
name = "quip-agent"
version = "1.0.0"
description = "Agent Service for Quip Bot"
authors = [
    {name = "Quip Team", email = "team@quip.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi",
    "pydantic",
    "httpx",
    "python-dotenv",
    "PyYAML",
    "langchain-openai",
    "langchain-core",
    "langchain (>=0.3.26,<0.4.0)",
    "langgraph",
    "langchain-mcp-adapters",
    "structlog",
    "redis>=5.0.0",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.21.0"
pytest-mock = ">=3.10.0"
black = ">=23.0.0"
isort = ">=5.12.0"
mypy = ">=1.5.0"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.21.0"
pytest-mock = ">=3.10.0"
pytest-cov = ">=4.0.0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
